requirements.txt
faker
fpdf2
Pillow
pymongo
pandas
pyyaml
boto3
